src/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dovvybuddy-backend"
version = "0.1.0"
description = "Python backend for DovvyBuddy (FastAPI + SQLAlchemy)"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.27.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.11.0",
    "google-generativeai>=0.3.2",
    "groq>=0.4.2",
    "tiktoken>=0.5.2",
    "tenacity>=8.2.3",
    "pgvector>=0.2.0",
    "jinja2>=3.1.3",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "scripts*"]
